(-) timestamp.py more cleanup
